requirements.txt
asgiref==3.11.0
attrs==25.4.0
boto3==1.42.25
botocore==1.42.25
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
uritemplate==4.2.0
urllib3==2.6.3
